requirements.txt
boto3==1.26.74
botocore==1.29.74
certifi==2022.12.7
charset-normalizer==3.0.1
environ==1.0
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
